Refactoring in ContextContainer and implementation (#10)
* Refactoring in ContextContainer

Rename SimpleContextContainer to DefaultContextContainer.

Drop ContextAccessorLoader and ThreadLocalAccessorLoader in favor of
configuring instances of those directly on DefaultContextContainer.
The layers above (e.g. Web Framework, GraphQL transport, etc.) have
the responsibility to know what accessors should be used, and therefore
ContextContainer should be directly configurable. In other words it
cannot be a global mechanism like ServiceLoader.

* Drop NOOP constants and implementations

If a ContextContainerPropagator isn't found, it should be an
IllegalStateException. It means we don't know how to save to a specific
external context, and something is wrong. Or otherwise it would
silently not propagate.

If a ContextContainer isn't present, there are several cases.
- the code expects a ContextContainer and therefore needs to check
  through the isNoop() method anyway to protect itself, but it might
  forget to do so.
- the code expects there may or may not be a ContextContainer, but
  wants to start a new one if not found. Here it also needs to check
  isNoop(), but if it doesn't do so, it could even lead to errors in
  code that expects an actual ContextContainer and not a noop.
- only code that expects there may or may not be a ContextContainer and
  doesn't care either way benefits; arguably though a noop lurking
  always has the potential of causing side effects and surprises.

For ContextContainerPropagator, we now always throw
IllegalStateException if there isn't a compatible one registered.

For ContextContainer, restoreFrom raises ISE while separately there
is a restoreIfPresentFrom that returns null if not found.

* Polishing and minor refactoring

Ordering of methods in DefaultContextContainer consistent with the
order in ContextContainer.

Rename ContextContainerPropagator to ContextContainerAdapter.