Refactor API and implementation
Replace ContextContainer with ContextRegistry to holds context accessors,
and ContextSnapshot that can be taken where context needs to propagate.

Update the accessor contracts to simplify implementations. For example
ThreadLocaAccessor is now really only concerned exposing the values but
not with maintaining any state.

There are additional features too such as restoring previous values for
thread locals.