Clearing ThreadLocals not present in source context
When a `Scope` is opened using `setAllThreadLocalsFrom`,
the values set for `ThreadLocalAccessor`s, that don't have
a corresponding key defined in the source context, should be cleared.
Upon closing the `Scope` these values should be restored.

Otherwise, the external values are leaking into the scope.

If a sum of `ThreadLocal`s and source context defined values needs
to be present in the opened `Scope`, the `captureAll` can be used instead.
A unit test is provided to show the difference in behaviour.