Unifying ContextSnapshot vocabulary and fixing setAllThreadLocalsFrom
Recent change to `ContextSnapshot#setThreadLocals()` method
allows to clear `ThreadLocal` values not present in the snapshot.

The associated change to `ContextSnapshot.setAllThreadLocalsFrom(sourceContext)`
should be on par with `ContextSnapshot.captureAll(...)` followed by
`ContextSnapshot#setThreadLocals()` but avoiding the temporary snapshot container.

Unfortunately, the method clears existing `ThreadLocal` values that
are not present in the sourceContext object and creates a mismatch in the vocabulary.

The previous behaviour is restored for that static method.

Instead, two new static methods are introduced:

- `ContextSnapshot.setThreadLocals(Object)`
- `ContextSnapshot.setThreadLocals(Object, ContextRegistry)`

that allow to skip the temporary container and clear `ThreadLocal`
values not present in the sourceContext provided and set only the
present ones.