ThreadLocalAccessor API capable of working in scoped scenarios
This change introduces two new methods: setValue(), and restore().

The first one can be called when entering a scope without a value defined,
while the second one allows clearing the ThreadLocal value upon closing a scope
when no surrounding ThreadLocal value was captured. This opens the possibility
to effectively clear ThreadLocal values if they are missing in the intended
scope (either coming from a ContextSnapshot or a Context object, acted upon by
the ContextAccessor).