ThreadLocalAccessor API capable of working in scoped scenarios (#103)
This change introduces two new methods: setValue(), and restore() to
ThreadLocalAccessor.  It also deprecates reset() in favour of the above.

setValue() can be called when entering a scope without a value defined, while
restore() allows clearing the ThreadLocal value upon closing a scope when no
surrounding ThreadLocal value was captured. This opens the possibility to
effectively clear ThreadLocal values if they are missing in the intended scope
(either coming from a ContextSnapshot or a Context object, acted upon by the
ContextAccessor).

---------

Co-authored-by: Marcin Grzejszczak <mgrzejszczak@vmware.com>