Optional ThreadLocalAccessor for SLF4J's MDC (#194)
Slf4jThreadLocalAccessor can be used together with SLF4J to propagate
the MDC across Thread boundaries.

This is an optional utility that can be used when SLF4J is on the
classpath. Additionally, the JSR305 nullability annotations are made
public in order to take advantage of them in the new package
(io.micrometer.context.integration).