Add more intuitive method to wrap ExecutorService
The `ContextExecutorService#wrap` method accepting a
`Supplier<ContextSnapshot>` can be a mystery for users as it's not
immediately obvious that a `ContextSnapshotFactory` needs to be used to
supply a `ContextSnapshot`. This change adds an overload that accepts
the `ContextSnapshotFactory` explicitly and only works with ThreadLocal
values as a sane default. For more advanced use cases the variant
accepting `Supplier` can be used.