Fix Slf4jThreadLocalAccessor with keys for missing values (#349)
In case of missing MDC entries, when a scope was closed, the value
should be removed. This was not the case and that could lead to leaks of
these MDC entries. This change takes the missing previous values into
account and clears them from the MDC in case they were set by a
snapshot.

Fixes #347.

Signed-off-by: Dariusz Jędrzejczyk <dariusz.jedrzejczyk@broadcom.com>