fix: Allow to remove a ThreadLocalAccessor by Key of Type Object
The `ThreadLocalAccessor.key()` is of Type Object. Thus, when it comes to delete a predefined `ThreadLocalAccessor` which have a non-String key, it's becomes impossible to delete.
This behavior was observed while attempting to remove `org.springframework.security.core.context.ReactiveSecurityContextHolderThreadLocalAccessor` from the `ContextRegistry`, which have a key of type `java.lang.Class`.

This change will just switch the method parameter in order to support the Type `java.lang.Object` as key, which shouldn't be a breaking change, even in a minor version.

Signed-off-by: Boubaker Khanfir <bkhanfir@exoplatform.com>